Removing unnecessary
